Thread Pool
Java Thread pool represents a group of worker thread that are waiting
for the job and reuse many times.

In case of thread pool a group of fixed sized thread are created. A
thread from the thread pool is pulled out and is assigned a job by the
service provider. After the completion of the job the thread is returned
back to the pool.

Advantage:

Better performance It saves time because there is no need to create new
thread.

Real time usage

It is used in Servlet and JSP where container creates a thread pool to
process the request.